# Cache results of flycheck-goto-line to speed up overlay creation