@cpitclaudel 2020-05-03
> Couldn't the cache be buffer-local?

Good point! And it saves the buffer check.